When renaming, refuse to overwrite an already existing entry
if an entry is attempted to be renamed check if an item with
same label, folder and wallet already exists and refuse to rename
if it does

BUG:458639

+++ src/runtime/kwalletd/secretserviceclient.cpp
@@ -819,6 +819,13 @@ void SecretServiceClient::renameEntry(const QString &display_name,
         return;
     }
 
+    SecretItemPtr existingItem = retrieveItem(newKey, folder, collectionName, ok);
+    if (existingItem) {
+        *ok = false;
+        qCWarning(KWALLETD_LOG) << i18n("Entry named %1 in folder %2 and wallet %3 already exists.", newKey, folder, collectionName);
+        return;
+    }
+
     deleteEntry(oldKey, folder, collectionName, ok);
     if (!*ok) {
         return;
